requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
github-copilot-sdk
